index controller cut out
